Fix users table ID datatype mismatch for foreign keys

+++ database/migrations/001_create_users_table.sql
@@ -1,7 +1,7 @@
 -- Slim auth-only users table for system operator accounts
 
 CREATE TABLE IF NOT EXISTS users (
-    id INT UNSIGNED NOT NULL AUTO_INCREMENT,
+    id BIGINT UNSIGNED NOT NULL AUTO_INCREMENT,
     name VARCHAR(255) NOT NULL,
     email VARCHAR(255) NOT NULL,
     password_hash VARCHAR(255) NOT NULL,
